Abstracting away common part for plugin and build file #SCL-21374
