+++ scala/scala-impl/resources/messages/ScalaBundle.properties
@@ -49,6 +49,7 @@ newclassorfile.menu.action.sbt.description=Create new Sbt File
 
 error.package.already.contains.build.sbt=''build.sbt'' already exists for package ''{0}''
 error.package.already.contains.plugin.sbt=''plugin.sbt'' already exists for package ''{0}''
+error.title.package.already.contains.file=Cannot Create File
 
 ### org/jetbrains/plugins/scala/actions/ShowTypeInfoAction.scala
 type.info.text=Type Info
